HW02/base step 3.

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,17 +1,20 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
     private final Map<Integer, MealDataBase> repository = new ConcurrentHashMap<>();
     private final AtomicInteger counter = new AtomicInteger(0);
@@ -34,21 +37,29 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        if (isBelongToUserId(repository.get(id).FOREIGN_KEY ,SecurityUtil.authUserId()))
+            return repository.remove(id) != null;
+        return false;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id).meal;
+        return isBelongToUserId(repository.get(id).FOREIGN_KEY ,SecurityUtil.authUserId())?repository.get(id).meal :null;
     }
 
     @Override
     public Collection<Meal> getAll() {
         return repository.values().stream()
+                .filter(mealDataBase -> isBelongToUserId(mealDataBase.FOREIGN_KEY ,SecurityUtil.authUserId()))
+                //.sorted(Comparator.comparing(mealDataBase -> mealDataBase.meal,(o, t1) -> t1.getDateTime().compareTo(o.getDateTime())))
                 .map(mealDataBase -> mealDataBase.meal)
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
                 .collect(Collectors.toList());
     }
 
+    private boolean isBelongToUserId(int foreignKey, int userId){
+        return foreignKey == userId;
+    }
     private static class MealDataBase {
         private final int FOREIGN_KEY = SecurityUtil.authUserId();
         private final Meal meal;
